// front-end/src/app/shared/enums/measurement-unit.enum.ts
export enum MeasurementUnit {
    UNITS = 'units',
    TEASPOON = 'teaspoon',
    TABLESPOON = 'tablespoon',
    CUP = 'cup',
    OUNCE = 'ounce',
    POUND = 'pound',
    GRAM = 'gram',
    KILOGRAM = 'kg',
    MILLILITER = 'ml',
    LITER = 'l',
    PIECE = 'piece',
    SLICE = 'slice',
    PINCH = 'pinch',
    WHOLE = 'whole'
}